+++ app/src/androidTest/java/com/mapbox/search/sample/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.mapbox.search.demo
+package com.mapbox.search.sample
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
