+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.mapbox.search.demo"
+    package="com.mapbox.search.sample"
     >
 
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
