+++ app/src/main/java/com/mapbox/search/sample/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.mapbox.search.demo
+package com.mapbox.search.sample
 
 import android.Manifest
 import android.content.Context
@@ -14,19 +14,19 @@ import android.widget.Toast
 import androidx.appcompat.widget.Toolbar
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
-import com.mapbox.search.demo.api.CategorySearchJavaExampleActivity
-import com.mapbox.search.demo.api.CategorySearchKotlinExampleActivity
-import com.mapbox.search.demo.api.ForwardGeocodingBatchResolvingJavaExampleActivity
-import com.mapbox.search.demo.api.ForwardGeocodingBatchResolvingKotlinExampleActivity
-import com.mapbox.search.demo.api.ForwardGeocodingJavaExampleActivity
-import com.mapbox.search.demo.api.ForwardGeocodingKotlinExampleActivity
-import com.mapbox.search.demo.api.OfflineReverseGeocodingJavaExampleActivity
-import com.mapbox.search.demo.api.OfflineReverseGeocodingKotlinExampleActivity
-import com.mapbox.search.demo.api.OfflineSearchJavaExampleActivity
-import com.mapbox.search.demo.api.OfflineSearchKotlinExampleActivity
-import com.mapbox.search.demo.api.ReverseGeocodingJavaExampleActivity
-import com.mapbox.search.demo.api.ReverseGeocodingKotlinExampleActivity
-import com.mapbox.search.demo.maps.MapsIntegrationExampleActivity
+import com.mapbox.search.sample.api.CategorySearchJavaExampleActivity
+import com.mapbox.search.sample.api.CategorySearchKotlinExampleActivity
+import com.mapbox.search.sample.api.ForwardGeocodingBatchResolvingJavaExampleActivity
+import com.mapbox.search.sample.api.ForwardGeocodingBatchResolvingKotlinExampleActivity
+import com.mapbox.search.sample.api.ForwardGeocodingJavaExampleActivity
+import com.mapbox.search.sample.api.ForwardGeocodingKotlinExampleActivity
+import com.mapbox.search.sample.api.OfflineReverseGeocodingJavaExampleActivity
+import com.mapbox.search.sample.api.OfflineReverseGeocodingKotlinExampleActivity
+import com.mapbox.search.sample.api.OfflineSearchJavaExampleActivity
+import com.mapbox.search.sample.api.OfflineSearchKotlinExampleActivity
+import com.mapbox.search.sample.api.ReverseGeocodingJavaExampleActivity
+import com.mapbox.search.sample.api.ReverseGeocodingKotlinExampleActivity
+import com.mapbox.search.sample.maps.MapsIntegrationExampleActivity
 import com.mapbox.search.result.SearchResult
 import com.mapbox.search.ui.view.SearchBottomSheetView
 import com.mapbox.search.ui.view.category.Category
