+++ app/src/main/java/com/mapbox/search/sample/api/ForwardGeocodingBatchResolvingKotlinExampleActivity.kt
@@ -1,4 +1,4 @@
-package com.mapbox.search.demo.api
+package com.mapbox.search.sample.api
 
 import android.app.Activity
 import android.os.Bundle
@@ -6,25 +6,26 @@ import android.util.Log
 import com.mapbox.search.MapboxSearchSdk
 import com.mapbox.search.ResponseInfo
 import com.mapbox.search.SearchEngine
+import com.mapbox.search.SearchMultipleSelectionCallback
 import com.mapbox.search.SearchOptions
 import com.mapbox.search.SearchRequestTask
 import com.mapbox.search.SearchSelectionCallback
 import com.mapbox.search.result.SearchResult
 import com.mapbox.search.result.SearchSuggestion
 
-class ForwardGeocodingKotlinExampleActivity : Activity() {
+class ForwardGeocodingBatchResolvingKotlinExampleActivity : Activity() {
 
     private lateinit var searchEngine: SearchEngine
     private lateinit var searchRequestTask: SearchRequestTask
 
-    private val searchCallback = object : SearchSelectionCallback {
+    private val searchCallback = object : SearchSelectionCallback, SearchMultipleSelectionCallback {
 
         override fun onSuggestions(suggestions: List<SearchSuggestion>, responseInfo: ResponseInfo) {
             if (suggestions.isEmpty()) {
                 Log.i("SearchApiExample", "No suggestions found")
             } else {
-                Log.i("SearchApiExample", "Search suggestions: $suggestions.\nSelecting first suggestion...")
-                searchRequestTask = searchEngine.select(suggestions.first(), this)
+                Log.i("SearchApiExample", "Search suggestions: $suggestions.")
+                searchRequestTask = searchEngine.select(suggestions, this)
             }
         }
 
@@ -44,6 +45,14 @@ class ForwardGeocodingKotlinExampleActivity : Activity() {
             Log.i("SearchApiExample", "Category search results: $results")
         }
 
+        override fun onResult(
+            suggestions: List<SearchSuggestion>,
+            results: List<SearchResult>,
+            responseInfo: ResponseInfo
+        ) {
+            Log.i("SearchApiExample", "Batch retrieve results: $results")
+        }
+
         override fun onError(e: Exception) {
             Log.i("SearchApiExample", "Search error", e)
         }
@@ -55,7 +64,7 @@ class ForwardGeocodingKotlinExampleActivity : Activity() {
 
         searchRequestTask = searchEngine.search(
             "Paris Eiffel Tower",
-            SearchOptions(limit = 5),
+            SearchOptions(),
             searchCallback
         )
     }
