+++ app/src/main/java/com/mapbox/search/sample/api/ForwardGeocodingJavaExampleActivity.java
@@ -1,4 +1,4 @@
-package com.mapbox.search.demo.api;
+package com.mapbox.search.sample.api;
 
 import android.os.Bundle;
 import android.util.Log;
@@ -10,7 +10,6 @@ import androidx.appcompat.app.AppCompatActivity;
 import com.mapbox.search.MapboxSearchSdk;
 import com.mapbox.search.ResponseInfo;
 import com.mapbox.search.SearchEngine;
-import com.mapbox.search.SearchMultipleSelectionCallback;
 import com.mapbox.search.SearchOptions;
 import com.mapbox.search.SearchRequestTask;
 import com.mapbox.search.SearchSelectionCallback;
@@ -19,7 +18,7 @@ import com.mapbox.search.result.SearchSuggestion;
 
 import java.util.List;
 
-public class ForwardGeocodingBatchResolvingJavaExampleActivity extends AppCompatActivity {
+public class ForwardGeocodingJavaExampleActivity extends AppCompatActivity {
 
     private SearchEngine searchEngine;
     private SearchRequestTask searchRequestTask;
@@ -31,8 +30,8 @@ public class ForwardGeocodingBatchResolvingJavaExampleActivity extends AppCompat
             if (suggestions.isEmpty()) {
                 Log.i("SearchApiExample", "No suggestions found");
             } else {
-                Log.i("SearchApiExample", "Search suggestions: " + suggestions);
-                searchRequestTask = searchEngine.select(suggestions, multipleSelection);
+                Log.i("SearchApiExample", "Search suggestions: " + suggestions + "\nSelecting first...");
+                searchRequestTask = searchEngine.select(suggestions.get(0), this);
             }
         }
 
@@ -52,19 +51,6 @@ public class ForwardGeocodingBatchResolvingJavaExampleActivity extends AppCompat
         }
     };
 
-    private final SearchMultipleSelectionCallback multipleSelection = new SearchMultipleSelectionCallback() {
-
-        @Override
-        public void onResult(@NonNull List<? extends SearchSuggestion> suggestions, @NonNull List<? extends SearchResult> results, @NonNull ResponseInfo responseInfo) {
-            Log.i("SearchApiExample", "Batch retrieve results: " + results);
-        }
-
-        @Override
-        public void onError(@NonNull Exception e) {
-            Log.i("SearchApiExample", "Search error: ", e);
-        }
-    };
-
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -72,6 +58,7 @@ public class ForwardGeocodingBatchResolvingJavaExampleActivity extends AppCompat
         searchEngine = MapboxSearchSdk.createSearchEngine();
 
         final SearchOptions options = new SearchOptions.Builder()
+            .limit(5)
             .build();
 
         searchRequestTask = searchEngine.search("Paris Eiffel Tower", options, searchCallback);
