+++ app/src/main/java/com/mapbox/search/sample/api/OfflineReverseGeocodingJavaExampleActivity.java
@@ -1,4 +1,4 @@
-package com.mapbox.search.demo.api;
+package com.mapbox.search.sample.api;
 
 import android.os.Bundle;
 import android.util.Log;
